chore: set memo phrase
