+++ src/pages/Home/index.tsx
@@ -27,8 +27,8 @@ function Home() {
   const [isDrag, setIsDrag] = useState(false);
   const [startX, setStartX] = useState(0);
   const [category, setCategory] = useState('all');
-  const [phrase, setPhrase] = useState(homePhraseData[0]);
 
+  const { postListData } = useGetPostListData(category, 0, 11);
   const onDragStart = (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
     setIsDrag(true);
@@ -56,32 +56,12 @@ function Home() {
   const changeCategory = (category: string) => {
     setCategory(category);
   };
-  const { postListData } = useGetPostListData(category, 0, 11);
-
-  const setPhraseData = useCallback(() => {
-    let index = 0;
-    setInterval(() => {
-      setPhrase(homePhraseData[index]);
-      index++;
-      if (index >= homePhraseData.length) index = 0;
-    }, 5000);
-  }, []);
-  useEffect(() => {
-    setPhraseData();
-  }, []);
-
   return (
     <>
       <HomeLayoutContainer>
         <HomeContentWrapper>
-          <HomePhraseWrapper
-            key={phrase.phrase}
-            animate={{ opacity: 1, y: 0 }}
-            initial={{ opacity: 0, y: 20 }}
-            exit={{ opacity: 1, y: -20 }}
-            transition={{ duration: 0.6 }}
-          >
-            <HomePhrase phraseData={phrase} />
+          <HomePhraseWrapper>
+            <HomePhrase />
           </HomePhraseWrapper>
           <CategoryMenu type={category} onClick={changeCategory} />
         </HomeContentWrapper>
